Moved more types from core to tailviewer
